add assertion
There's a double check for accordance of the task type.

One of the checks must not be executed.

Perform assertion within the AddTaskCommand class, as it's guaranteed
that the task already has an appropriate type.

Assertion is used as reaching this line clearly indicates inconsistency
in code and in logic.